Chore: Added controls to ContextMenu story (#54373)
