+++ packages/grafana-ui/src/components/ContextMenu/WithContextMenu.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 
 import { ContextMenu } from '../ContextMenu/ContextMenu';
 
-interface WithContextMenuProps {
+export interface WithContextMenuProps {
   /** Menu item trigger that accepts openMenu prop */
   children: (props: { openMenu: React.MouseEventHandler<HTMLElement> }) => JSX.Element;
   /** A function that returns an array of menu items */
